% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler.R
\name{collecter}
\alias{collecter}
\title{DataFrame Collecter Function}
\usage{
collecter(df, sample = FALSE, method = "fraction", size = 1,
  replace = FALSE, seed = NULL)
}
\arguments{
\item{df}{Spark dataframe}

\item{sample}{logical option to apply sampling prior to collecting data into
R. Default is FALSE which applies no sampling}

\item{method}{sampling method to apply. Valid options include "fraction" and
"head" for both R and Spark and additionally "n" and "tail" for R}

\item{size}{size of sampling. fraction of data for "fraction" method. number
of records for other methods}

\item{replace}{logical option to allow for record replacement. valid option
for "fraction" and "n" methods}

\item{seed}{option numerical input to set random seed. Default is NULL which
does not set the random seed and results will vary}
}
\value{
R data.frame
}
\description{
Collect Spark Dataframe into local R with options to sample data prior to
collecting.
}
\details{
If sample option selected, arguments passed to sampler function for sampling
prior to collecting
}
