% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexer.R
\name{date_parter}
\alias{date_parter}
\alias{date_parter.data.frame}
\alias{date_parter.tbl_spark}
\title{Dataframe Date Parter Funtion}
\usage{
date_parter(df, measure_vars, label, abbr, ...)

\method{date_parter}{data.frame}(df, measure_vars, label = FALSE,
  abbr = FALSE, ...)

\method{date_parter}{tbl_spark}(df, measure_vars, ...)
}
\arguments{
\item{df}{data.frame}

\item{measure_vars}{column name of Date class to calculate date_parts on. Can
be more than 1 column}

\item{label}{logical option to change month and day of week values to label.
default is FALSE}

\item{abbr}{logical option to use abbreviate labels. default is FALSE}

\item{...}{not currently implemented}
}
\value{
data.frame
}
\description{
Function to calculate and append date part values to a data.frame
}
\details{
Calculates following date parts
   year    - numerical year
   quarter - numerical quarter
   month   - numerical month value. can be change to month label for R method
   day_of_month - numerical day of month - day_of_week('2000-01-01') = 1
   day_of_week - numerical weekday. 1 = Sunday. can be changed to label for R method
   day_of_year - numerical day of the year - day_of_year('2000-01-06') = 6
   days_left_in_month - number of days left in current month - days_left_in_month('2000-12-31') = 0
}
\examples{
df <- data.frame(today = Sys.Date())
date_parter(df, "today")
}
