% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateconverter.R
\name{dateconverter}
\alias{dateconverter}
\alias{dateconverter.data.frame}
\alias{dateconverter.tbl_spark}
\title{Date Converter Function}
\usage{
dateconverter(df, input_col_name, input_format)

\method{dateconverter}{data.frame}(df, input_col_name, input_format)

\method{dateconverter}{tbl_spark}(df, input_col_name, input_format)
}
\arguments{
\item{df}{DataFrame}

\item{input_col_name}{Input string column which is to be converted to date}

\item{input_format}{Input format of String column. Allowed formats are
MM/dd/yyyy HH:mm:ss, MM/dd/yyyy, yyyy/MM/dd HH:mm:ss, yyyy/MM/dd,
dd/MM/yyyy HH:mm:ss, dd/MM/yyyy. The forward slash symbol can be interchanged
with hifen.}
}
\value{
returns DataFrame with converted Date column added with name as
{INPUT_COLUMN_NAME}_CONV & format as either yyyy-MM-dd or yyyy-MM-dd HH:mm:ss.
}
\description{
Date Converter function which is used to convert String date column with
given format to Date or Datetime field with format as either yyyy-MM-dd 
or yyyy-MM-dd HH:mm:ss
}
\examples{

library(dplyr)
library(sparklyr)
library(lubridate)

dateconverter(df, "EVENT_TIMESTAMP", "MM/dd/yyyy HH:mm:ss", "yyyy-MM-dd HH:mm:ss", "month", "start")
}
