% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateformatter.R
\name{dateformatter}
\alias{dateformatter}
\alias{dateformatter.data.frame}
\alias{dateformatter.tbl_spark}
\title{Date Period Formatter Function}
\usage{
dateformatter(df, input_col_name, input_format, output_format)

\method{dateformatter}{data.frame}(df, input_col_name, input_format,
  output_format = "yyyy-MM-dd HH:mm:ss")

\method{dateformatter}{tbl_spark}(df, input_col_name, input_format,
  output_format = "yyyy-MM-dd HH:mm:ss")
}
\arguments{
\item{df}{DataFrame}

\item{input_col_name}{Input Date column on which transformation has to be applied}

\item{input_format}{Input format of Date column. Allowed formats are
MM/dd/yyyy HH:mm:ss, MM/dd/yyyy, yyyy/MM/dd HH:mm:ss, yyyy/MM/dd,
dd/MM/yyyy HH:mm:ss, dd/MM/yyyy. The forward slash symbol can be interchanged
with hifen.}

\item{output_format}{Required format for the output Date column. Default value
is "yyyy-MM-dd HH:mm:ss"}
}
\value{
returns DataFrame with formatted string column added with name as
{INPUT_COLUMN_NAME}_FMT
}
\description{
Date Period formatter function which is configurable & can be used in
other a2 functions to get required formatted Date value. Return column
will be of the class string.
}
\examples{

library(dplyr)
library(sparklyr)
library(lubridate)

dateformatter(df, "10/14/2018 23:59:59", "MM/dd/yyyy HH:mm:ss", "yyyy-MM-dd HH:mm:ss")
Return value = "2018-10-14 23:59:59"
}
