% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_schema}
\alias{get_schema}
\title{Get Dataset Schema}
\usage{
get_schema(df)
}
\arguments{
\item{df}{dataset}
}
\value{
named list of column types
}
\description{
Returns named list of column types. Names refer to column names. Uses schema
function to apply correct method
}
\examples{
get_schema(mtcars)
}
