% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputer.R
\name{impute_constant}
\alias{impute_constant}
\title{Impute Constant}
\usage{
impute_constant(df, measure_vars = NULL, fill = 0)
}
\arguments{
\item{df}{Spark or R data.frame}

\item{measure_vars}{optional character vector of column names to apply impute
function on. default is NULL which applies impute on all columns with
missing values}

\item{fill}{constant value to fill missing values with}
}
\value{
imputed Spark or R data.frame
}
\description{
Function to impute missing values with constant value provided
}
\details{
The fill type will determine which columns the imputation is applied on.
Setting fill to 5 (a numeric type) will only impute missing values for
numberic columns
}
