% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputer.R
\name{impute_constant_at}
\alias{impute_constant_at}
\title{Impute Numeric Constant At}
\usage{
impute_constant_at(df, measure_vars, fill)
}
\description{
Function imputes numeric constants for specific columns
}
