% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputer.R
\name{imputer}
\alias{imputer}
\title{Dataframe Imputer}
\usage{
imputer(df, group_vars = NULL, measure_vars = NULL, fun = "mean", ...)
}
\arguments{
\item{df}{Spark or R data.frame}

\item{group_vars}{optional grouping variable input. Only valid for mean and
mode impute functions. default is NULL which applies no grouping. note -
mode function does not supporting grouping}

\item{measure_vars}{optional character vector of column names to apply impute
function on. default is NULL which applies impute on all columns with
missing values}

\item{fun}{impute function. valid options are mean, mode, or constant.
default is mean}

\item{...}{optional additional arguments to pass to impute function. ex-
constant requires a fill input}
}
\value{
imputed Spark or R data.frame
}
\description{
Function to impute missing values from either Spark or R data.frames.
}
\details{
Imputer fills missing values by either calculating the average value, most
frequent or some user defined constant
}
