% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexer.R
\name{indexer}
\alias{indexer}
\alias{indexer.data.frame}
\alias{indexer.tbl_spark}
\title{Dataframe Index Function}
\usage{
indexer(df, measure_vars, origin, units, periods, label)

\method{indexer}{data.frame}(df, measure_vars, origin = "2000-01-01",
  units = "days", periods = 1, label = NULL)

\method{indexer}{tbl_spark}(df, measure_vars, origin = "2000-01-01",
  units = "days", periods = 1, label = NULL)
}
\arguments{
\item{df}{data.frame}

\item{measure_vars}{column name of Date class to calculate index on. Can be
more than 1 column}

\item{origin}{either string that can be converted to a Date or column name of
seperate Date class}

\item{units}{time interval to measure time difference on. Valid units are
years, months, weeks, days, hours, minutes, and seconds for R method and
days and months for Spark}

\item{periods}{number of time periods to count. Valid only for R method}

\item{label}{optional input for custom column name of output}
}
\value{
data.frame
}
\description{
Function calculates the time between two date values as a numerical index.
}
\details{
Spark version only caculates index for days and months units and single
period
}
\examples{
df <- data.frame(today = Sys.Date())
indexer(df, "today", origin = "2018-05-20", units = "days", periods = 1)
indexer(df, "today", origin = "2018-05-20", units = "hours", periods = 1)
indexer(df, "today", origin = "2018-05-20", units = "hours", periods = 4)
}
