% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputer.R
\name{na_cols}
\alias{na_cols}
\title{Get NA Column names from dataframe}
\usage{
na_cols(df)
}
\arguments{
\item{df}{dataframe}
}
\value{
vector of column names
}
\description{
Function to get the column names for any columns with missing values.
Function works for both Spark and R dataframes
}
