% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler.R
\name{sampler}
\alias{sampler}
\alias{sampler.data.frame}
\alias{sampler.tbl_spark}
\title{DataFrame Sampler Generic Function}
\usage{
sampler(df, group_vars = NULL, method, size, replace, weight = NULL,
  seed)

\method{sampler}{data.frame}(df, group_vars = NULL,
  method = "fraction", size = 1, replace = FALSE, weight = NULL,
  seed = NULL)

\method{sampler}{tbl_spark}(df, group_vars = NULL, method = "fraction",
  size = 1, replace = FALSE, weight = NULL, seed = NULL)
}
\arguments{
\item{df}{DataFrame}

\item{group_vars}{optional grouping variable. Allows for stratified sampling
for R data.frames only. default is NULL which applies no grouping}

\item{method}{sampling method to apply. Valid options include "fraction" and
"head" for both R and Spark and additionally "n" and "tail" for R}

\item{size}{size of sampling. fraction of data for "fraction" method. number
of records for other methods}

\item{replace}{logical option to allow for record replacement. valid option
for "fraction" and "n" methods}

\item{weight}{optional option to apply sampling weights. see
\url{https://dplyr.tidyverse.org/reference/sample.html} for more details.
Default is NULL for no weighting. Only valid for R data.frames}

\item{seed}{option numerical input to set random seed. Default is NULL which
does not set the random seed and results will vary}
}
\value{
Dataframe
}
\description{
Function to sample dataset. Supports multiple methods for both spark and R
data.frames
}
\details{
Grouping and weighting options only supporting in base R. Fraction and head
sampling methods valid for both R and Spark while n and tail only valid for R
}
