% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{schema_check}
\alias{schema_check}
\title{Schema Check}
\usage{
schema_check(x, y, validate = TRUE)
}
\value{
data.frame with schema comparison if stop error not thrown
}
\description{
Function to check to if a schema matches a second. Validate option will
create a stop error if not all of x columns present in y or if there is a
type mismatch
}
\details{
Function supports either dataframe inputs (R or Spark) or schema lists
}
\examples{

# Schema Input Example
schema_check(get_schema(mtcars), get_schema(select(mtcars, mpg, am)))

# Dataframe Input Example
schema_check(mtcars, select(mtcars, mpg, am))
}
