% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{sim_df}
\alias{sim_df}
\title{Simulated Behavioral Dataset}
\format{A tibble with 2000 rows and 7 variables 
 \describe{
  \item{index}{integer type with sequential index of records}
  \item{id}{integer type. example of an unique identifier}
  \item{date}{date type. date of record}
  \item{cat1}{character type. record attribute example. 2 distinct values.}
  \item{cat2}{character type. record attribute example. 5 disitinct values.}
  \item{metric1}{integer type. numerical metric example. 5 distinct values.}
  \item{metric2}{numeric type. numerical metric example.}
  }}
\usage{
sim_df
}
\description{
A simulated dataset to represent a behavioral dataset with date and id based
records
}
\keyword{datasets}
