% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{sim_df_anom}
\alias{sim_df_anom}
\title{Simulated Time Series with Anomalies Dataset}
\format{A tibble with 100 rows and 4 variables 
 \describe{
  \item{index}{integer type with sequential index of records}
  \item{date}{date type. date of record}
  \item{y}{numeric type. time series value. created with arima.sim function. anomalies added}
  \item{flag}{integer type. flag for anomalies added. 1 = positive anomaly, -1 = negative anomaly, 0 = no anomaly }
  }}
\usage{
sim_df_anom
}
\description{
A simulated time series dataset with anomalies to be used in detecter testing
}
\keyword{datasets}
