% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roller.R
\name{sql_fun_translator}
\alias{sql_fun_translator}
\title{Aggregation Function SQL Translation}
\usage{
sql_fun_translator(fun)
}
\description{
Helper function to translate native r function to spark sql syntax
}
