% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateformatter.R
\name{dateformatter}
\alias{dateformatter}
\alias{dateformatter.data.frame}
\alias{dateformatter.tbl_spark}
\title{Date Period Formatter Function}
\usage{
dateformatter(df, measure_vars, input_format, output_format, output_suffix)

\method{dateformatter}{data.frame}(df, measure_vars, input_format,
  output_format = "yyyy-MM-dd HH:mm:ss", output_suffix = "FMT")

\method{dateformatter}{tbl_spark}(df, measure_vars, input_format,
  output_format = "yyyy-MM-dd HH:mm:ss", output_suffix = "FMT")
}
\arguments{
\item{df}{DataFrame}

\item{measure_vars}{Input Date column on which transformation has to be applied}

\item{input_format}{Input format of Date column. Allowed formats are
MM/dd/yyyy HH:mm:ss, MM/dd/yyyy, yyyy/MM/dd HH:mm:ss, yyyy/MM/dd,
dd/MM/yyyy HH:mm:ss, dd/MM/yyyy. The forward slash symbol can be interchanged
with hifen.}

\item{output_format}{Required format for the output Date column. Default value
is "yyyy-MM-dd HH:mm:ss"}

\item{output_suffix}{Suffix to be added to the new column generated by the function}
}
\value{
returns DataFrame with formatted string column added with name as
{INPUT_COLUMN_NAME}_FMT
}
\description{
Date Period formatter function which is configurable & can be used in
other a2 functions to get required formatted Date value. Return column
will be of the class string.
}
\examples{

library(dplyr)
library(lubridate)

date_func_df <- data.frame(TIME_COL = as.POSIXlt(c("2017-01-01 10:15:15", "2017-09-23 14:26:59", "2017-11-15 05:05:05", "2018-05-11 08:15:18", "2018-03-27 23:59:59")), stringsAsFactors = FALSE)

dateformatter(date_func_df, "TIME_COL", "yyyy-MM-dd HH:mm:ss", "MM/dd/yyyy HH:mm:ss")
}
